did a bunch of stuff -- smaller commit next time
