+++ Cargo.toml
@@ -9,7 +9,9 @@ path = "src/main.rs"
 
 [dependencies]
 anyhow = "1.0.98"
+chacha20poly1305 = "0.10.1"
 clap = { version = "4.5.36", features = ["derive"] }
+rand_core = "0.9.3"
 serial_test = "3.2.0"
 sled = "0.34.7"
 tempfile = "3.19.1"
