+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod cli;
+pub mod storage;
+pub mod encryption;
